# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  companyId: string;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface Company {
  id: string;
  name: string;
}

export interface CreateCompanyRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  name: string;
}

export interface CompanyEnvelope {
  company: Company;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
  /** @minLength 1 */
  companyId?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  field_rep: 'field_rep',
  manager: 'manager',
  admin: 'admin',
} as const;

export type WorkflowAssignment = typeof WorkflowAssignment[keyof typeof WorkflowAssignment];


export const WorkflowAssignment = {
  retail: 'retail',
  insurance: 'insurance',
  both: 'both',
} as const;

export type PinWorkflow = typeof PinWorkflow[keyof typeof PinWorkflow];


export const PinWorkflow = {
  retail: 'retail',
  insurance: 'insurance',
} as const;

export type DamageType = typeof DamageType[keyof typeof DamageType];


export const DamageType = {
  roof: 'roof',
  siding: 'siding',
  roof_and_siding: 'roof_and_siding',
} as const;

export type DoorKnockResult = typeof DoorKnockResult[keyof typeof DoorKnockResult];


export const DoorKnockResult = {
  no_answer: 'no_answer',
  no_appointment: 'no_appointment',
  appointment: 'appointment',
} as const;

export type ContactOutcome = typeof ContactOutcome[keyof typeof ContactOutcome];


export const ContactOutcome = {
  no_soliciting: 'no_soliciting',
  priority_inspection: 'priority_inspection',
  call_to_schedule: 'call_to_schedule',
} as const;

export interface Profile {
  userId: string;
  role: Role;
  workflowAssignment: WorkflowAssignment;
}

export interface ProfileEnvelope {
  profile: Profile;
}

export interface RetailData {
  /** @minLength 1 */
  ownerName1: string;
  /** @nullable */
  ownerName2?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  interestedRoof: boolean;
  interestedSiding: boolean;
  interestedWindows: boolean;
  interestedDoors: boolean;
  /** @nullable */
  interestNotes?: string | null;
  /** @nullable */
  appointmentDate?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface Pin {
  id: string;
  userId: string;
  latitude: number;
  longitude: number;
  /** @nullable */
  address: string | null;
  workflow: PinWorkflow;
  damageType: DamageType | null;
  /** @nullable */
  photoUrl: string | null;
  doorKnockResult: DoorKnockResult | null;
  retailData: RetailData | null;
  contactOutcome: ContactOutcome | null;
  /** @nullable */
  customerName: string | null;
  /** @nullable */
  customerPhone: string | null;
  status: string;
  createdAt: string;
}

export interface PinEnvelope {
  pin: Pin;
}

export interface PinListEnvelope {
  pins: Pin[];
}

export interface CreatePinInput {
  latitude: number;
  longitude: number;
  workflow: PinWorkflow;
  damageType?: DamageType;
  photoUrl?: string;
  doorKnockResult?: DoorKnockResult;
  retailData?: RetailData;
  contactOutcome?: ContactOutcome;
  customerName?: string;
  customerPhone?: string;
}

export interface BulkPinInput {
  latitude: number;
  longitude: number;
  /** @minLength 1 */
  photoUrl: string;
}

export interface BulkCreatePinsInput {
  /** @minItems 1 */
  pins: BulkPinInput[];
}

export const DeleteSuccessValue = {
  success: true,
} as const;
export type DeleteSuccess = typeof DeleteSuccessValue;

export interface AdminStats {
  totalPins: number;
  insurancePins: number;
  retailPins: number;
  appointments: number;
  fieldRepCount: number;
}

export interface AdminStatsEnvelope {
  stats: AdminStats;
}

export interface TeamUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  role: Role;
  workflowAssignment: WorkflowAssignment;
  pinCount: number;
  joinedAt: string;
}

export interface TeamUserEnvelope {
  user: TeamUser;
}

export interface TeamUserListEnvelope {
  users: TeamUser[];
}

export interface UpdateTeamUserInput {
  role?: Role;
  workflowAssignment?: WorkflowAssignment;
}

export interface LocationPingBody {
  latitude: number;
  longitude: number;
}

export const LocationPingSuccessValue = {
  success: true,
} as const;
export type LocationPingSuccess = typeof LocationPingSuccessValue;

export interface TeamLocation {
  userId: string;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  latitude: number;
  longitude: number;
  updatedAt: string;
}

export interface TeamLocationListEnvelope {
  locations: TeamLocation[];
}

export interface ReverseGeocodeResponse {
  /** @nullable */
  address: string | null;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
/**
 * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
 */
returnTo?: string;
/**
 * Company ID chosen on the join/create-company screen. Only used the first time this user logs in; ignored on subsequent logins.
 */
companyId?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};

export type ListPinsParams = {
/**
 * Managers/admins only — filter to a single rep's pins.
 */
userId?: string;
};

export type ReverseGeocodeCoordinatesParams = {
latitude: number;
longitude: number;
};