# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stats:
    get:
      tags: [Admin]
      operationId: getAdminStats
      summary: Get overview KPI stats
      description: Manager/admin only.
      responses:
        "200":
          description: Overview stats.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatsEnvelope"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Manager/admin role required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/users:
    get:
      tags: [Admin]
      operationId: listTeamUsers
      summary: List all users with role, workflow assignment, and pin counts
      description: Manager/admin only.
      responses:
        "200":
          description: Team member list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamUserListEnvelope"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Manager/admin role required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/users/{userId}:
    patch:
      tags: [Admin]
      operationId: updateTeamUser
      summary: Update a team member's role and/or workflow assignment
      description: |
        Manager/admin only. Enforces role-hierarchy permission rules: nobody
        can modify their own role, managers cannot modify other managers or
        admins and cannot grant manager/admin, admins can manage everyone.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamUserInput"
      responses:
        "200":
          description: Updated team member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamUserEnvelope"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not permitted to perform this update.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [Admin]
      operationId: removeTeamUser
      summary: Remove a team member
      description: Manager/admin only, subject to the same role-hierarchy rules as update.
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSuccess"
        "401":
          description: Not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Not permitted to remove this user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: User not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Role:
      type: string
      enum: [field_rep, manager, admin]
    WorkflowAssignment:
      type: string
      enum: [retail, insurance, both]
...
    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    AdminStats:
      type: object
      required:
        [
          totalPins,
          insurancePins,
          retailPins,
          appointments,
          fieldRepCount,
        ]
      properties:
        totalPins:
          type: integer
        insurancePins:
          type: integer
        retailPins:
          type: integer
        appointments:
          type: integer
        fieldRepCount:
          type: integer
    AdminStatsEnvelope:
      type: object
      required: [stats]
      properties:
        stats:
          $ref: "#/components/schemas/AdminStats"
    TeamUser:
      type: object
      required:
        [
          id,
          email,
          firstName,
          lastName,
          profileImageUrl,
          role,
          workflowAssignment,
          pinCount,
          joinedAt,
        ]
      properties:
        id:
          type: string
        email:
          type: [string, "null"]
        firstName:
          type: [string, "null"]
        lastName:
          type: [string, "null"]
        profileImageUrl:
          type: [string, "null"]
        role:
          $ref: "#/components/schemas/Role"
        workflowAssignment:
          $ref: "#/components/schemas/WorkflowAssignment"
        pinCount:
          type: integer
        joinedAt:
          type: string
          format: date-time
    TeamUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/TeamUser"
    TeamUserListEnvelope:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/TeamUser"
    UpdateTeamUserInput:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/Role"
        workflowAssignment:
          $ref: "#/components/schemas/WorkflowAssignment"
...